modules: battery: channel definition into per-module header
Separating battery channel definition into dedicated per-module header.
