+++ app/src/common/message_channel.c
@@ -42,14 +42,6 @@ ZBUS_CHAN_DEFINE(NETWORK_CHAN,
 		 ZBUS_MSG_INIT(.type = NETWORK_DISCONNECTED)
 );
 
-ZBUS_CHAN_DEFINE(BATTERY_CHAN,
-		 struct battery_msg,
-		 NULL,
-		 NULL,
-		 ZBUS_OBSERVERS_EMPTY,
-		 ZBUS_MSG_INIT(0)
-);
-
 ZBUS_CHAN_DEFINE(ERROR_CHAN,
 		 enum error_type,
 		 NULL,
