+++ app/src/modules/app/app.c
@@ -13,6 +13,7 @@
 
 #include "modules_common.h"
 #include "message_channel.h"
+#include "battery.h"
 
 /* Register log module */
 LOG_MODULE_REGISTER(app, CONFIG_APP_LOG_LEVEL);
