+++ app/src/modules/battery/CMakeLists.txt
@@ -5,3 +5,4 @@
 #
 
 target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/battery.c)
+target_include_directories(app PRIVATE .)
