+++ tests/module/button/src/main.c
@@ -10,6 +10,7 @@
 #include <zephyr/logging/log.h>
 
 #include "message_channel.h"
+#include "battery.h"
 
 #include <dk_buttons_and_leds.h>
 #include <date_time.h>
