+++ tests/module/cloud/CMakeLists.txt
@@ -22,6 +22,7 @@ zephyr_include_directories(${ZEPHYR_BASE}/include/zephyr/)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/testsuite/include)
 zephyr_include_directories(../../../app/src/modules/cloud)
 zephyr_include_directories(../../../app/src/common)
+zephyr_include_directories(../../../app/src/modules/battery)
 zephyr_include_directories(${NRF_DIR}/subsys/net/lib/nrf_cloud/include)
 zephyr_include_directories(${NRF_DIR}/../modules/lib/cjson)
